Moved logic for indexing to analyzer
