+++ ressie/ressie/queries/query.py
@@ -1,34 +1,22 @@
 from __future__ import print_function
 
 import datetime
-import os
 
-import whoosh.index as index
 from elasticsearch import Elasticsearch
-from whoosh.qparser import QueryParser
 
+from ressie.analyzer.http import Http
 from ressie.models import Hit
-from ..analyzer.http import Http
 
 
 class ElasticQuery(object):
     # in minutes
     time_threshold = 200000
-    index_folder = os.getcwd() + "/data/index/"
 
     def __init__(self):
         pass
 
     def check_status(self):
         self.elasticsearch()
-
-    def check_attack_db(self, attack):
-        ix = index.open_dir(self.index_folder)
-        qp = QueryParser("lovrotestira", schema=ix.schema)
-        q = qp.parse(u"%s" % attack)
-
-        with ix.searcher() as s:
-            s.search(q, limit=20)
 
     def elasticsearch(self):
